Moving BundleHelper and PersonHelper to their own files
