+++ src/blueink/__init__.py
@@ -2,6 +2,6 @@ from requests import exceptions
 
 from . import constants
 from .client import Client
-from .model.bundles import BundleHelper
+from .bundle_helper import BundleHelper
 
 __all__ = ['Client', 'BundleHelper', 'exceptions', 'constants']
