+++ src/blueink/tokenizedrequests.py
@@ -1,13 +1,8 @@
-import io
 import json
 
 import requests
 
 from munch import munchify
-
-"""
-Helper functions to add header with user's token.txt
-"""
 
 
 class Pagination:
